% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendAlnCat.R
\name{appendAlnCat}
\alias{appendAlnCat}
\title{appendAlnCat}
\usage{
appendAlnCat(database, alnID, directory)
}
\arguments{
\item{database}{A database that will have the alignment ID and the
corresponding category appended to it.}

\item{alnID}{The alignment ID as a string.}

\item{directory}{The directory which alnID is contained in as a string.}
}
\value{
The database with the alignment category appended to the
alnCategory data frame contained in the database.
}
\description{
Append the alignment ID with the corresponding reference
ID it belongs to into the provided database.
}
\keyword{internal}
