% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendMatrix.R
\name{appendMatrix}
\alias{appendMatrix}
\title{appendMatrix.R}
\usage{
appendMatrix(database, readaln, alnID)
}
\arguments{
\item{database}{A database that will have the alignment in matrix form
appended to it.}

\item{readaln}{An object of class alignment produced by seqinr. It contains
the number of aligned sequences, a vector of strings containing the names
of the aligned sequences, a vector of strings containing the aligned
sequence, and a vector of strings containing commentaries for each sequence
or NA is there is no comments.}

\item{alnID}{The alignment ID as a character vector.}
}
\value{
The database with the alignment matrix appended to the alignments
list contained in the database.
}
\description{
\code{<function>} Append the alignment to the database provided in the
form of a matrix. Rownames are set as the sequence names in the alignment.
Each row contains the corresponding sequence and each column is the
individual characters of the sequences.
}
\details{
Details.
}
\keyword{internal}
