% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendSeqCat.R
\name{appendSeqCat}
\alias{appendSeqCat}
\title{appendSeqCat}
\usage{
appendSeqCat(database, readaln, alnID)
}
\arguments{
\item{database}{A database that will have the alignment in matrix form
appended to it.}

\item{readaln}{An object of class alignment produced by seqinr. It contains
the number of aligned sequences, a vector of strings containing the names
of the aligned sequences, a vector of strings containing the aligned
sequence, and a vector of strings containing commentaries for each sequence
or \code{NA} is there is no comments.}

\item{alnID}{The alignment ID as a string.}
}
\value{
The database with the sequence ID and it's corresponding alignment
ID appended to the data frame seqCategory contained in the the database.
}
\description{
\code{<function>} Append the sequence ID and it's corresponding alignment
ID to the provided database.
}
\keyword{internal}
