% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertXStringSet.R
\name{convertXStringSet}
\alias{convertXStringSet}
\title{convertXStringSet}
\usage{
convertXStringSet(filePath, readType = c("AA", "DNA", "RNA"))
}
\arguments{
\item{filePath}{The path to the file of sequences stored in a .mfa
file in the form of a character vector.}

\item{readType}{The type of characters found in the provided input
sequences. Can either be "AA", "DNA" or "RNA" depending on the characters
in the file.}
}
\value{
A \code{XStringSet} object corresponding to the \code{readType}
input.
}
\description{
\code{<function>} The convertXStringSet function converts a provided
.mfa file into a XStringSet corresponding to the \code{readType} provided.
If the provided readType is invalid, \code{convertXStringSet} will stop
all running functions and prompt the user to input a valid readType.
}
\details{
Details.
}
\keyword{internal}
