% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDB.R
\name{createDB}
\alias{createDB}
\title{createDB}
\usage{
createDB()
}
\value{
A data model containing the alignments and sequences of
the BaliBASE database.
}
\description{
\code{<function>} Produce a datamodel containing the contents from the
BaliBASE database. Refer to the data model to understand the structure
of the reference data base.
}
\keyword{internal}
