% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchAln.R
\name{fetchAln}
\alias{fetchAln}
\title{fetchAln.R}
\usage{
fetchAln(alnID, collapse = TRUE, asMatrix = TRUE)
}
\arguments{
\item{alnID}{A character vector of the alignment name.}

\item{collapse}{TRUE if the user would like to remove gaps in the sequence.
FALSE if they would like to keep the gaps. @collapse is automatically
set to TRUE.}

\item{asMatrix}{TRUE if the user would like the sequence in matrix form.
FALSE will return a character vector of the sequence. @asMatrix is
automatically set to FALSE.}

\item{seq}{A character vector of the sequence name in quotations.}
}
\value{
The alignment of interest from the BaliBASE database.
}
\description{
\code{<function>} Return an alignment of interest from BaliBASE.
}
\details{
Details.
}
\section{Input}{
 The alignment ID of the alignment of interest from BaliBASE.
Additonal inputs are collapse and asMatrix which are automatically
set to FALSE if not provided.
}

