% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchMuscle.R
\name{fetchMuscle}
\alias{fetchMuscle}
\title{fetchMuscle.R}
\usage{
fetchMuscle(inputSeqs, readType = c("AA", "DNA", "RNA"),
  cluster = "default", gapOpening = "default", gapExtension = "default",
  maxiters = "default", substitutionMatrix = "default", type = "default",
  order = c("aligned", "input"), verbose = FALSE, help = FALSE, ...)
}
\arguments{
\item{inputSeqs}{The sequences the user would like to align. Can either
be an alignment ID of a BaliBASE alignment or a file path to a .mfa file
containing sequences for an alignment.}

\item{readType}{The type of characters found in the provided input
sequences. Can either be "AA", "DNA" or "RNA" depending on the characters
in the file.

For the remaining parameters:
Please view details on msa::msa() function which provides details about the
remaining parameters for writeMuscle}
}
\value{
The alignment generated by Muscle in the form of a seqinr
alignment.
}
\description{
\code{<function>} Return a seqinr alignment containing sequences that are
aligned by the Muscle alignment algorithm. The user may search the
referenceDB for a set of BaliBASE sequences by entering the alignment ID
as input, or they may input a .mfa file containing the sequences in
fasta format.
}
\details{
Details.
}
