% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchSeq.R
\name{fetchSeq}
\alias{fetchSeq}
\title{fetchSeq}
\usage{
fetchSeq(seqName, collapse = TRUE, asMatrix = FALSE)
}
\arguments{
\item{seqName}{A character vector of the sequence name in quotations.}

\item{collapse}{TRUE if the user would like to remove gaps in the sequence.
FALSE if they would like to keep the gaps. \code{collapse} is automatically
set to TRUE.}

\item{asMatrix}{TRUE if the user would like the sequence in matrix form.
FALSE will return a character vector of the sequence. \code{asMatrix} is
automatically set to FALSE.}
}
\value{
The sequence of interest from the BaliBASE database.
}
\description{
Return a sequence of interest from BaliBASE collpased
or with gaps and/or in matrix format.
}
\section{Input}{
 The name of the sequence of interest as a character vector.
}

\examples{
\dontrun{
seqID <- "1aab_"

seq <- fetchSeq(seqID, collapse = FALSE)
seq2 <- fetchSeq(seqID, asMatrix = FALSE)
}

}
