% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlnID.R
\name{getAlnID}
\alias{getAlnID}
\title{getAlnID}
\usage{
getAlnID(referenceCat = NULL, seqID = NULL)
}
\arguments{
\item{referenceCat}{A reference category in the form of a character
vector. \code{getAlnID} will return all alignment ID's that are
contained in that reference category.}

\item{seqID}{A sequence ID in the form of a character vector.
\code{getAlnID} will return the alignment ID's containing the
specified sequence.}
}
\value{
Either a list of alnID's that are contained in the
specified reference category or a list of alnID's that contain
the specified sequence ID's.
}
\description{
\code{<function>} Retrieve alignment ID's in a corresponding reference
category of BaliBASE or retrieve the alignment ID's corresponding
to a specific sequence ID.
}
