% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeqID.R
\name{getSeqID}
\alias{getSeqID}
\title{getSeqID}
\usage{
getSeqID(alnID)
}
\arguments{
\item{alnID}{An alignment ID in the form of a character vector.}
}
\value{
A list of sequence ID's are containedin the specified
alignment ID.
}
\description{
Retrieve sequence ID's that are contained in the
alignment ID provided
}
\examples{
\dontrun{
seqIds <- getSeqID("BB11001")
(seqIds)
}

}
