% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameFile.R
\name{nameFile}
\alias{nameFile}
\title{nameFile.R}
\usage{
nameFile(inputID = NULL, fileName, extension, directory)
}
\arguments{
\item{inputID}{The name of the file as a string if fileName is NULL.
Automatically set to NULL.}

\item{fileName}{The name of the file if the user provides a desired
fileName. Automatically set to NULL.}

\item{extension}{The extension of the file provided by the user.}
}
\value{
The generated file name.
}
\description{
\code{<function>} The @nameFile function generates a name for a file. If the
passed fileName is NULL, then concatenate the @inputID and @extension.
If not, then concatenate the passed @fileName and @extension and return
the final name of the file.
}
\details{
Details.
}
\keyword{internal}
