% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referenceDB.R
\docType{data}
\name{referenceDB}
\alias{referenceDB}
\title{referenceDB is a data set utilized in AQE.}
\description{
referenceDB is a data set utilized in AQE.
}
\references{
\url{https://github.com/judyheewonlee/AQE}
}
\author{
Heewon Lee \email{heewon.lee@mail.utoronto.ca}
}
\keyword{data}
