% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreAlignment.R
\name{scoreAlignment}
\alias{scoreAlignment}
\title{scoreAlignment.R}
\usage{
scoreAlignment(reference, test, SP = FALSE, CS = FALSE)
}
\arguments{
\item{reference}{The filepath as a character vector to the reference
alignment.}

\item{test}{The pilepath as a character vector to the alignment that will
be compared with the reference alignment.}

\item{SP}{A boolean value, TRUE if the user would like the sum of pairs
score included into the returned PAC object. Automatically set to
FALSE.}

\item{CS}{A boolean value, TRUE if the user would like the total column
score included into the returned PAC object. Automatically set to
FALSE.}
}
\value{
A PAC object containing providing the optimal pairwise column
alignment of two alternative MSAs of the same sequences, and summary
statistics of the differences between them. Refer to the documentation
for @compare_alignments function for details on the output components
in the AlignStat package.
}
\description{
\code{<function>} The @scoreAlignment function takes a reference alignment
and a test alignment and returns an object of class "pairwise alignment
comparison" (PAC) containing the sum of pairs and/or total column scores
using the AlignStat package.
}
\details{
Details.
}
