% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeClustalW.R
\name{writeClustalW}
\alias{writeClustalW}
\title{writeClustalW}
\usage{
writeClustalW(inputSeqs, readType = c("AA", "DNA", "RNA"), fileName = NULL,
  directory = "inst/extdata/Output", cluster = "default",
  gapOpening = "default", gapExtension = "default", maxiters = "default",
  substitutionMatrix = "default", type = "default", order = c("aligned",
  "input"), verbose = FALSE, help = FALSE, ...)
}
\arguments{
\item{inputSeqs}{The sequences the user would like to align. Can either
be an alignment ID of a BaliBASE alignment or a file path to a .mfa file
containing sequences for an alignment.}

\item{readType}{The type of characters found in the provided input
sequences. Can either be "AA", "DNA" or "RNA" depending on the characters
in the file.}

\item{fileName}{The name of the ClustalW alignment file as a character
vector. Automatically set to NULL, where the file will be named as the
concatenated string of the \code{alnID} and "ClustalW".}

\item{directory}{The directory the user would like the file to be placed in.
Must be a string. Note: "/" should not be added to the end of the directory.}

\item{cluster}{The clustering method which should be used. Possible values
are \code{"nj"} (default) and \code{"upgma"}. In the original ClustalW
implementation, this parameter is called clustering.}

\item{gapOpening}{gap opening penalty; the default value for nucleotide
sequences is 15.0, the default value for amino acid sequences is 10.0.}

\item{gapExtension}{gap extension penalty; the default value for
nucleotide sequences is 6.66, the default value for amino acid
sequences is 0.2.}

\item{maxiters}{maximum number of iterations; the default value
is 16. In the original ClustalW implementation, this parameter
is called \code{numiters}.}

\item{substitutionMatrix}{substitution matrix for scoring matches and
mismatches; can be a real matrix, a file name, or the name of a built-in
substitution matrix. In the latter case, the choices \code{"blosum"},
\code{"pam"}, \code{"gonnet"}, and \code{"id"} are supported
for amino acid sequences. For aligning nucleotide sequences,
the choices \code{"iub"} and \code{"clustalw"} are possible.
The parameter dnamatrix can also be used instead for the sake of backwards
compatibility. The valid choices for this parameter are \code{"iub"} and
\code{"clustalw"}. In the original ClustalW implementation, this parameter
is called \code{matrix}.}

\item{type}{type of the input sequences \code{inputSeqs}; see \code{msa()}
from \code{msa} package.}

\item{order}{how the sequences should be ordered in the output object
(see msa); in the original ClustalW implementation, this parameter is
called \code{outorder}.}

\item{verbose}{if TRUE, the algorithm displays detailed information
and progress messages.}

\item{help}{if TRUE, information about algorithm-specific parameters is
displayed. In this case, no multiple sequence alignment is performed and
the function quits after displaying the additional help information.}

\item{...}{further parameters specific to ClustalW; An overview of parameters
that are available in this interface is shown when calling \code{msaClustalW}
with \code{help=TRUE}. For more details, see also the documentation of
ClustalW.}
}
\description{
Write a multifasta file containing sequences that are
aligned by the ClustalW alignment algorithm. The user may search the
referenceDB for a set of BaliBASE sequences by entering the alignment ID
as input, or they may input a .mfa file containing the sequences in
fasta format.
}
\examples{
\dontrun{
writeClustalW("BB11001", readType = "AA")

writeAln("BB11002")
writeClustalW("data/Output/BB11002.mfa",
                        readType = "AA", gapExtension = 2)
}

}
\seealso{
Please view details on \code{\link[msa]{msa}} and
\code{\link[msa]{msaClustalW}} function which provides details about
the parameters that adjust the alignment.
}
