% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSeq.R
\name{writeSeq}
\alias{writeSeq}
\title{writeSeq}
\usage{
writeSeq(seqID, fileName = NULL, directory = "inst/extdata/Output",
  collapse = TRUE)
}
\arguments{
\item{seqID}{The Balibase seq ID in the form of a character vector.
Note: Be sure to provide the correct cases as \code{writeSeq} is
case sensitive.}

\item{fileName}{The name of the alignment file as a character vector.
Automatically set to NULL, where the file will be named as the \code{seqID}.}

\item{directory}{The directory the user would like the file to be placed in.
Must be a character vector. Note: "/" should not be added to the end of the
directory.}

\item{collapse}{TRUE if the user would like to collapse the gaps in the
alignment. FALSE to maintain gaps. \code{collapse} is automatically set to
TRUE.}
}
\description{
The writeSeq function returns a sequence from the
database generated from the Balibase suite.
}
\examples{
\dontrun{
seq <- "1ton_"
fileName <- "1ton_seq"
writeSeq(seq, fileName)

seq <- "1aab_"
writeSeq(seq, collapse = FALSE)
}

}
