% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameFile.R
\name{nameFile}
\alias{nameFile}
\title{nameFile}
\usage{
nameFile(inputID = NULL, fileName, extension, directory)
}
\arguments{
\item{inputID}{The name of the file as a string if fileName is NULL.
Automatically set to NULL.}

\item{fileName}{The name of the file if the user provides a desired
fileName. Automatically set to NULL.}

\item{extension}{The extension of the file provided by the user.}
}
\value{
The generated file name.
}
\description{
The \code{nameFile} function generates a name for a
file. If the passed fileName is NULL, then concatenate the
\code{inputID} and \code{extension.} If not, then concatenate the
passed \code{fileName} and \code{extension} and return
the final name of the file.
}
\keyword{internal}
