% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeMuscle.R
\name{writeMuscle}
\alias{writeMuscle}
\title{writeMuscle.R}
\usage{
writeMuscle(inputSeqs, readType = c("AA", "DNA", "RNA"), fileName = NULL,
  directory = "data/Output", cluster = "default", gapOpening = "default",
  gapExtension = "default", maxiters = "default",
  substitutionMatrix = "default", type = "default", order = c("aligned",
  "input"), verbose = FALSE, help = FALSE, ...)
}
\arguments{
\item{inputSeqs}{The sequences the user would like to align. Can either
be an alignment ID of a BaliBASE alignment or a file path to a .mfa file
containing sequences for an alignment.}

\item{readType}{The type of characters found in the provided input
sequences. Can either be "AA", "DNA" or "RNA" depending on the characters
in the file.}

\item{fileName}{The name of the Muscle alignment file as a character vector.
Automatically set to NULL, where the file will be named as the concatenated
string of the @alnID and "Muscle".}

\item{directory}{The directory the user would like the file to be placed in.
Must be a string. Note: "/" should not be added to the end of the directory.

For the remaining parameters:
Please view details on msa::msa() function which provides details about the
remaining parameters for writeMuscle}
}
\description{
\code{<function>} Write a multifasta file containing sequences that are
aligned by the ClustalW alignment algorithm. The user may search the
referenceDB for a set of BaliBASE sequences by entering the alignment ID
as input, or they may input a .mfa file containing the sequences in
fasta format.
}
\details{
Details.
}
